+++ backend/sokdak/src/test/java/com/wooteco/sokdak/member/service/EmailServiceTest.java
@@ -5,7 +5,6 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import com.wooteco.sokdak.auth.service.AuthCodeGenerator;
 import com.wooteco.sokdak.member.dto.EmailRequest;
 import com.wooteco.sokdak.member.repository.TicketRepository;
 import com.wooteco.sokdak.ticket.domain.Ticket;
@@ -13,7 +12,6 @@ import com.wooteco.sokdak.util.ServiceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
 
 class EmailServiceTest extends ServiceTest {
 
@@ -22,12 +20,6 @@ class EmailServiceTest extends ServiceTest {
 
     @Autowired
     private TicketRepository ticketRepository;
-
-    @MockBean
-    private AuthCodeGenerator authCodeGenerator;
-
-    @MockBean
-    private EmailSender emailSender;
 
     @DisplayName("인증 코드 발송")
     @Test
